Fixed Hunter Summary DVC/Facing Count decimal point bug

+++ fetch_details_hunter.php
@@ -89,7 +89,7 @@ for ($i = 0; $i < count($hunter_summary); $i++){
         if ($brand_count == NULL) {
             $brand_count = 0;
         }
-        $hunter_summary[$i]["Brand Hunted"] += (int)($brand_count * 1.5 * $weight);
+        $hunter_summary[$i]["Brand Hunted"] += (float)($brand_count * 1.5 * $weight);
         $sql = 'SELECT COUNT(*) FROM '.$dbname.'.products a WHERE a.product_type ="sku" AND a.account_id = :account_id AND (a.product_creation_time >= :start_datetime AND a.product_creation_time <= :end_datetime) AND a.product_status = 2';
         $stmt = $pdo->prepare($sql);
         $stmt->execute(['account_id'=>$hunter_summary[$i][probe_processed_hunter_id], 'start_datetime'=>strval($_POST['start_datetime']), 'end_datetime'=>strval($_POST['end_datetime'])]);
@@ -97,7 +97,7 @@ for ($i = 0; $i < count($hunter_summary); $i++){
         if ($sku_count == NULL) {
             $sku_count = 0;
         }
-        $hunter_summary[$i]["SKU Hunted"] += (int)($sku_count * 1 * $weight );
+        $hunter_summary[$i]["SKU Hunted"] += (float)($sku_count * 1 * $weight );
         $sql = 'SELECT COUNT(*) FROM '.$dbname.'.products a WHERE product_type ="dvc" AND a.account_id = :account_id AND (a.product_creation_time >= :start_datetime AND a.product_creation_time <= :end_datetime) AND a.product_status = 2';
         $stmt = $pdo->prepare($sql);
         $stmt->execute(['account_id'=>$hunter_summary[$i][probe_processed_hunter_id], 'start_datetime'=>strval($_POST['start_datetime']), 'end_datetime'=>strval($_POST['end_datetime'])]);
@@ -105,7 +105,7 @@ for ($i = 0; $i < count($hunter_summary); $i++){
         if ($dvc_count == NULL) {
             $dvc_count = 0;
         }
-        $hunter_summary[$i]["DVC Hunted"] += (int)($dvc_count * 0.5 *$weight);
+        $hunter_summary[$i]["DVC Hunted"] += (float)($dvc_count * 0.5 *$weight);
         $sql = 'SELECT SUM(a.product_facing_count) FROM '.$dbname.'.products a WHERE a.account_id = :account_id AND (a.product_creation_time >= :start_datetime AND a.product_creation_time <= :end_datetime) AND a.product_status = 2';
         $stmt = $pdo->prepare($sql);
         $stmt->execute(['account_id'=>$hunter_summary[$i][probe_processed_hunter_id], 'start_datetime'=>strval($_POST['start_datetime']), 'end_datetime'=>strval($_POST['end_datetime'])]);
@@ -113,7 +113,7 @@ for ($i = 0; $i < count($hunter_summary); $i++){
         if ($facing_count == NULL) {
             $facing_count = 0;
         }
-        $hunter_summary[$i]["Hunted Facing Count"] += (int)($facing_count* 0.5 *$weight);
+        $hunter_summary[$i]["Hunted Facing Count"] += (float)($facing_count* 0.5 *$weight);
         $sql = 'SELECT COUNT(*) FROM '.$dbname.'.probe a WHERE a.probe_processed_hunter_id = :account_id AND (a.probe_hunter_processed_time >= :start_datetime AND a.probe_hunter_processed_time <= :end_datetime)';
         $stmt = $pdo->prepare($sql);
         $stmt->execute(['account_id'=>$hunter_summary[$i][probe_processed_hunter_id], 'start_datetime'=>strval($_POST['start_datetime']), 'end_datetime'=>strval($_POST['end_datetime'])]);
